Refactor #coerce to use Coercer
This one was actually really pretty freakin' cool. Ruby has built-in
methods (defined on Object) that coerce a lot of types for you (e.g.
Integer, String, Float, etc.). We took advantage of that and defined a
few more (and redefined some) so that by passing in the class name, you
can coerce a type using the pre-defined methods.